# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Tree API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/trees/{tree_id}": {
      "get": {
        "tags": [
          "trees"
        ],
        "summary": "Get Tree",
        "operationId": "get_tree_api_trees__tree_id__get",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "trees"
        ],
        "summary": "Update Tree",
        "operationId": "update_tree_api_trees__tree_id__patch",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trees"
        ],
        "summary": "Delete Tree",
        "operationId": "delete_tree_api_trees__tree_id__delete",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trees/{tree_id}/people/{person_id}/events": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Add Person Event",
        "operationId": "add_person_event_api_trees__tree_id__people__person_id__events_post",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/families/{family_id}/events": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Add Family Event",
        "operationId": "add_family_event_api_trees__tree_id__families__family_id__events_post",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "family_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Family Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/events/{event_id}": {
      "patch": {
        "tags": [
          "events"
        ],
        "summary": "Update Event",
        "operationId": "update_event_api_trees__tree_id__events__event_id__patch",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "events"
        ],
        "summary": "Delete Event",
        "operationId": "delete_event_api_trees__tree_id__events__event_id__delete",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/families/{family_id}": {
      "patch": {
        "tags": [
          "events"
        ],
        "summary": "Update Family",
        "operationId": "update_family_api_trees__tree_id__families__family_id__patch",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "family_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Family Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/places": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Places",
        "operationId": "list_places_api_trees__tree_id__places_get",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceOut"
                  },
                  "title": "Response List Places Api Trees  Tree Id  Places Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DateQualifier": {
        "type": "string",
        "enum": [
          "exact",
          "about",
          "before",
          "after",
          "between"
        ],
        "title": "DateQualifier"
      },
      "EventIn": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "default": ""
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "title": "Description",
            "default": ""
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FuzzyDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "place": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "EventIn"
      },
      "EventOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FuzzyDateOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          }
        },
        "type": "object",
        "required": [
          "id",
          "person_id",
          "family_id",
          "type",
          "title",
          "description",
          "date",
          "place"
        ],
        "title": "EventOut"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "birth",
          "baptism",
          "death",
          "burial",
          "education",
          "occupation",
          "retirement",
          "residence",
          "emigration",
          "immigration",
          "military",
          "engagement",
          "marriage",
          "separation",
          "divorce",
          "other"
        ],
        "title": "EventType"
      },
      "EventUpdate": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventType"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FuzzyDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "place": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          }
        },
        "type": "object",
        "title": "EventUpdate"
      },
      "FamilyOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/PartnerStatus"
          },
          "partner_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Partner Ids"
          },
          "child_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Child Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "partner_ids",
          "child_ids"
        ],
        "title": "FamilyOut"
      },
      "FamilyUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/PartnerStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FamilyUpdate"
      },
...
      "FuzzyDate": {
        "properties": {
          "qualifier": {
            "$ref": "#/components/schemas/DateQualifier",
            "default": "exact"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "year2": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year2"
          },
          "month2": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Month2"
          },
          "day2": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day2"
          },
          "phrase": {
            "type": "string",
            "maxLength": 200,
            "title": "Phrase",
            "default": ""
          }
        },
        "type": "object",
        "title": "FuzzyDate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyAccessOut": {
        "properties": {
          "tree_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "subtree_roles": {
            "items": {
              "$ref": "#/components/schemas/SubtreeRoleOut"
            },
            "type": "array",
            "title": "Subtree Roles"
          },
          "highest_role": {
            "$ref": "#/components/schemas/Role"
          },
          "my_person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Person Id"
          },
          "can_manage_members": {
            "type": "boolean",
            "title": "Can Manage Members"
          },
          "can_manage_subtrees": {
            "type": "boolean",
            "title": "Can Manage Subtrees"
          },
          "can_create_people": {
            "type": "boolean",
            "title": "Can Create People"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          }
        },
        "type": "object",
        "required": [
          "tree_role",
          "subtree_roles",
          "highest_role",
          "my_person_id",
          "can_manage_members",
          "can_manage_subtrees",
          "can_create_people",
          "is_owner"
        ],
        "title": "MyAccessOut"
      },
      "PartnerStatus": {
        "type": "string",
        "enum": [
          "together",
          "separated",
          "divorced"
        ],
        "title": "PartnerStatus"
      },
...
      "PlaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "PlaceOut"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "personal",
          "contributor",
          "admin",
          "owner"
        ],
        "title": "Role",
        "description": "Membership roles, lowest to highest privilege."
      },
...
      "SubtreeRoleOut": {
        "properties": {
          "subtree_id": {
            "type": "string",
            "format": "uuid",
            "title": "Subtree Id"
          },
          "subtree_name": {
            "type": "string",
            "title": "Subtree Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "subtree_id",
          "subtree_name",
          "role"
        ],
        "title": "SubtreeRoleOut"
      },
...
      "TreeDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "access": {
            "$ref": "#/components/schemas/MyAccessOut"
          },
          "person_count": {
            "type": "integer",
            "title": "Person Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "access",
          "person_count"
        ],
        "title": "TreeDetailOut"
      },
...
      "TreeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at"
        ],
        "title": "TreeOut"
      },
      "TreeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "TreeUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}